+++ tests/test_train_ms2query_model.py
@@ -8,10 +8,8 @@ from ms2query.create_new_library.train_ms2query_model import (
     DataCollectorForTraining, calculate_tanimoto_scores_with_library,
     convert_to_onnx_model, train_ms2query_model, train_random_forest)
 from ms2query.ms2library import MS2Library
-from ms2query.utils import (load_matchms_spectrum_objects_from_file,
-                            load_ms2query_model, load_pickled_file,
-                            predict_onnx_model)
-
+from ms2query.utils import predict_onnx_model
+from matchms import Spectrum
 
 if sys.version_info < (3, 8):
     pass
@@ -19,56 +17,33 @@ else:
     pass
 
 
-@pytest.fixture
-def path_to_test_dir():
-    return os.path.join(
-        os.path.split(os.path.dirname(__file__))[0],
-        'tests/test_files')
-
-
-@pytest.fixture
-def ms2library(path_to_test_dir):
-    path_to_general_tests_dir = os.path.join(path_to_test_dir, 'general_test_files')
-
-    return MS2Library(sqlite_file_name=os.path.join(path_to_general_tests_dir, "100_test_spectra.sqlite"),
-                      s2v_model_file_name=os.path.join(path_to_general_tests_dir, "100_test_spectra_s2v_model.model"),
-                      ms2ds_model_file_name=os.path.join(path_to_general_tests_dir,
-                                                         "ms2ds_siamese_210301_5000_500_400.hdf5"),
-                      pickled_s2v_embeddings_file_name=os.path.join(path_to_general_tests_dir,
-                                                                    "100_test_spectra_s2v_embeddings.pickle"),
-                      pickled_ms2ds_embeddings_file_name=os.path.join(path_to_general_tests_dir,
-                                                                      "100_test_spectra_ms2ds_embeddings.pickle"),
-                      ms2query_model_file_name=None)
-
-
-@pytest.fixture
-def query_spectra(path_to_test_dir):
-    training_spectra_file_name = os.path.join(
-        path_to_test_dir,
-        "test_files_train_ms2query_nn/20_training_spectra.mgf")
-    return load_matchms_spectrum_objects_from_file(training_spectra_file_name)
-
-
 def test_data_collector_for_training_init(ms2library):
     """Tests if an object DataCollectorForTraining can be created"""
     DataCollectorForTraining(ms2library)
 
 
-def test_get_matches_info_and_tanimoto(tmp_path, ms2library, query_spectra):
-    select_data_for_training = DataCollectorForTraining(ms2library)
-    result = select_data_for_training.get_matches_info_and_tanimoto(query_spectra)
-    expected_result = load_pickled_file(os.path.join(
-        os.path.split(os.path.dirname(__file__))[0],
-        "tests/test_files/test_files_train_ms2query_nn",
-        "expected_train_and_val_data.pickle"))[:2]
-    assert isinstance(result, tuple), "Expected tuple to be returned"
-    assert len(result) == 2, "Expected tuple to be returned"
-    pd.testing.assert_frame_equal(result[0], expected_result[0], check_dtype=False, check_exact=False, rtol=1e-1)
-    pd.testing.assert_frame_equal(result[1], expected_result[1], check_dtype=False, check_exact=False, rtol=1e-1)
+def test_get_matches_info_and_tanimoto(ms2library, hundred_test_spectra):
+    preselection_cut_off = 2
+    select_data_for_training = DataCollectorForTraining(ms2library, preselection_cut_off=preselection_cut_off)
+    training_scores, training_labels = select_data_for_training.get_matches_info_and_tanimoto(hundred_test_spectra)
+    assert isinstance(training_scores, pd.DataFrame)
+    assert isinstance(training_labels, pd.DataFrame)
+    assert training_scores.shape == (preselection_cut_off * len(hundred_test_spectra), 5)
+    assert training_labels.shape == (preselection_cut_off * len(hundred_test_spectra), 1)
+    assert list(training_scores.columns) == ['precursor_mz_library_spectrum',
+                                             'precursor_mz_difference',
+                                             's2v_score',
+                                             'average_ms2deepscore_multiple_library_structures',
+                                             'average_tanimoto_score_library_structures']
+    assert list(training_labels.columns) == ['Tanimoto_score']
+    assert round(training_scores.loc[0, "average_tanimoto_score_library_structures"], ndigits=5) == 0.57879
 
 
-def test_calculate_all_tanimoto_scores(tmp_path, ms2library, query_spectra):
-    query_spectrum = query_spectra[0]
+def test_calculate_all_tanimoto_scores(tmp_path, ms2library):
+    query_spectrum = Spectrum(mz=np.array([], dtype="float"),
+                              intensities=np.array([], dtype="float"),
+                              metadata={"smiles": "CC1=CC(=O)O[C@@H](CCC[C@@H](CCCC(CCC[C@@H](CCC[C@H](CC(C[C@@H](CCCC(CCC[C@H](C[C@H](CCCCC1)O)O)O)O)O)O)O)O)O)C(C)(C)C"},
+                              metadata_harmonization=False)
     spectra_ids_list = \
         [38, 3, 60]
     result = calculate_tanimoto_scores_with_library(ms2library.sqlite_library, query_spectrum, spectra_ids_list)
@@ -79,11 +54,10 @@ def test_calculate_all_tanimoto_scores(tmp_path, ms2library, query_spectra):
     pd.testing.assert_frame_equal(result, expected_result, check_dtype=False)
 
 
-def test_train_and_save_random_forest():
-    training_scores, training_labels = load_pickled_file(os.path.join(
-        os.path.split(os.path.dirname(__file__))[0],
-        "tests/test_files/test_files_train_ms2query_nn",
-        "expected_train_and_val_data.pickle"))[:2]
+def test_train_and_save_random_forest(ms2library, hundred_test_spectra):
+    select_data_for_training = DataCollectorForTraining(ms2library, preselection_cut_off=1)
+    training_scores, training_labels = select_data_for_training.get_matches_info_and_tanimoto(hundred_test_spectra)
+
     ms2query_model = train_random_forest(training_scores, training_labels)
     onnx_model = convert_to_onnx_model(ms2query_model)
     onnx_model_session = InferenceSession(onnx_model.SerializeToString())