+++ tests/test_utils.py
@@ -14,17 +14,6 @@ def test_convert_files_to_matchms_spectrum_objects_unknown_file(tmp_path):
         load_matchms_spectrum_objects_from_file(os.path.join(tmp_path, "file_that_does_not_exist.json"))
 
 
-def test_convert_files_to_matchms_spectrum_object_known_file():
-    """Test if pickled file is loaded in correctly"""
-    spectra = load_matchms_spectrum_objects_from_file(os.path.join(
-        os.path.split(os.path.dirname(__file__))[0],
-        'tests/test_files/general_test_files/100_test_spectra.mgf'))
-    assert isinstance(spectra, list), "expected list of spectra"
-    assert len(spectra) == 100, "expected 100 spectra"
-    for spectrum in spectra:
-        assert isinstance(spectrum, Spectrum)
-
-
 def test_add_unknown_charges_to_spectra(hundred_test_spectra):
     spectra = hundred_test_spectra
     # Set charges to predefined values